src:  Fix missing From<std::io::Error> impl when compiling with memfs enabled but localfs disabled
